Refactor db.js location
