+++ controllers/hashtags.js
@@ -1,4 +1,4 @@
-const { db, parseSqlUpdateStmt } = require('../config/db');
+const { db, parseSqlUpdateStmt } = require('../db/db');
 const asyncHandler = require('../middleware/async');
 const ErrorResponse = require('../utils/errorResponse');
 const { cleanseData } = require('../utils/dbHelper');
