+++ middleware/auth.js
@@ -1,7 +1,7 @@
 const jwt = require('jsonwebtoken');
 const asyncHandler = require('./async');
 const ErrorResponse = require('../utils/errorResponse');
-const { db } = require('../config/db');
+const { db } = require('../db/db');
 
 // Protect routes
 exports.protect = asyncHandler(async (req, res, next) => {
