+++ server.js
@@ -5,7 +5,7 @@ const colors = require('colors');
 const cookieParser = require('cookie-parser');
 const morgan = require('morgan');
 
-const { checkConn } = require('./config/db');
+const { checkConn } = require('./db/db');
 const apiRoutes = require('./routes/api-routes');
 const testRoutes = require('./routes/test-routes');
 const errorHandler = require('./middleware/error');
